test: 정적 리소스 반환용 컨트롤러 테스트 분리
